+++ src/m17/m17/__main__.py
@@ -1,4 +1,6 @@
-helptext = """
+from m17.core.constants import DEFAULT_TEST_HOST
+
+helptext = f"""
 `python -m m17` doesn't do anything on it's own (yet).
 Examples:
 
@@ -10,7 +12,7 @@ With `pip install m17[Codec2]`
     python -m m17.audio_test 3200
 
     Use your microphone mute as a reverse push to talk with these:
-    python -m m17.bareclient m17tester.tarxvf.tech 17010 tx srccall dstcall
+    python -m m17.bareclient {DEFAULT_TEST_HOST} 17010 tx srccall dstcall
     python -m m17.bareclient localhost 17000 rx
     python -m m17.bareclient raspi.lan 17000 full srccall dstcall